fix(api): add validation
