+++ package.json
@@ -76,6 +76,7 @@
   },
   "dependencies": {
     "@algolia/dns-filter": "1.1.17",
+    "altheia-async-data-validator": "5.0.11",
     "body-parser": "1.19.0",
     "cookie-parser": "1.4.5",
     "csurf": "1.11.0",
