+++ src/api/constants.ts
@@ -10,3 +10,8 @@ export const CSP_HEADERS = [
   'img-src * data:',
   'font-src *',
 ].join('; ');
+
+export const SESSION_COOKIE = 'sessionToken=53cu23_535510n; SameSite=Strict';
+
+export const DELETE_COOKIE =
+  'sessionToken=; expires=Thu, 01 Jan 1970 00:00:00 GMT; SameSite=Strict';
